♻️ 서비스워커 등록 및 권한 요청 코드를 파일 분리함 (#19)
